[docs] Fixing broken link in what books were migrated (#1494)
* [docs] Fixing broken link in what books were migrated

The modified link was incorrect. I have updated it to the main page of the `asciidocalypse` repository.

But I have a concern... we say:

> A full list of the content that was migrated to the V3 format is available here

But in the asciidocalypse we don't have such a list (or I can't find it), but the leftovers, if I'm not mistaken.

Where is the actual list that we should be offering?

* Update docs/migration/guide/mapping.md

---------

Co-authored-by: Jan Calanog <jan.calanog@elastic.co>