Remove Uri sources from navigation rendering. (#1546)
* Remove Uri sources from navigation rendering.

This PR removes the need to keep track of NavigationSource on markdown files to render the navigation.

Previous to this we needed the keep this around to track the 'real' parent of a section once it may have been reshuffled by `navigation.yml`.

We now more diligently build the global navigation more truthfully and can relay on walking the current navigation items `.Parent` to find the proper root to render.

This includes scrubbing/skipping phantoms from the navigation.

Another step to refactoring our navigation acyclic graph as the single source of truth.

* remove temporary pragma's while refactoring

* use logger not Console.WriteLine