Add `repo init-config` command
This allows the build to source the configuration prior to cloning and building

docs-assembler repo init-config
docs-assembler repo clone-all
docs-assembler repo build-all

On production we can lock the configuration to a commit using `--git-ref sha`.

ConfigurationFileProvider will look at

- {pwd}/config, development
- {app_data}/config, (where init-config clones too)
- embedded sources

We also now log the configuration source and in case of app_data also the git reference.